+++ tests/checkers/test_matching_ast.py
@@ -1,14 +1,40 @@
-def test_check_captures_w200(check_text):
-    res = check_text(
-        """\
-        def foo():
-            if bar():
-                return baz(quux("snork"))
-            else:
-                return baz(quux("snork"))
-        """,
-        filename="foo.py",
-    )
+import pytest
+
+
+@pytest.mark.parametrize("expr", ("if", "try", "ifexpr", "ifexpr-in-if-test"))
+def test_check_captures_w200(check_text, expr):
+    if expr == "if":
+        content = """\
+            def foo():
+                if bar():
+                    return baz(quux("snork"))
+                else:
+                    return baz(quux("snork"))
+        """
+    elif expr == "try":
+        content = """\
+            def foo():
+                try:
+                    return baz(quux("snork"))
+                except:
+                    return baz(quux("snork"))
+        """
+    elif expr == "ifexpr":
+        content = """\
+            def foo():
+                return baz(quux("snork")) if buzz() else baz(quux("snork"))
+        """
+    elif expr == "ifexpr-in-if-test":
+        content = """\
+            def foo():
+                if baz(quux("snork")) if buzz() else baz(quux("snork")):
+                    return 1
+                return 0
+        """
+    else:
+        raise NotImplementedError
+
+    res = check_text(content, filename="foo.py")
 
     assert not res.success
     assert (